+++ Tests/LinuxMain.swift
@@ -714,6 +714,12 @@ extension ModifierOrderTests {
     ]
 }
 
+extension MultilineArgumentsBracketsRuleTests {
+    static var allTests: [(String, (MultilineArgumentsBracketsRuleTests) -> () throws -> Void)] = [
+        ("testWithDefaultConfiguration", testWithDefaultConfiguration)
+    ]
+}
+
 extension MultilineArgumentsRuleTests {
     static var allTests: [(String, (MultilineArgumentsRuleTests) -> () throws -> Void)] = [
         ("testMultilineArgumentsWithDefaultConfiguration", testMultilineArgumentsWithDefaultConfiguration),
@@ -725,6 +731,18 @@ extension MultilineArgumentsRuleTests {
 
 extension MultilineFunctionChainsRuleTests {
     static var allTests: [(String, (MultilineFunctionChainsRuleTests) -> () throws -> Void)] = [
+        ("testWithDefaultConfiguration", testWithDefaultConfiguration)
+    ]
+}
+
+extension MultilineLiteralBracketsRuleTests {
+    static var allTests: [(String, (MultilineLiteralBracketsRuleTests) -> () throws -> Void)] = [
+        ("testWithDefaultConfiguration", testWithDefaultConfiguration)
+    ]
+}
+
+extension MultilineParametersBracketsRuleTests {
+    static var allTests: [(String, (MultilineParametersBracketsRuleTests) -> () throws -> Void)] = [
         ("testWithDefaultConfiguration", testWithDefaultConfiguration)
     ]
 }
@@ -1400,8 +1418,11 @@ XCTMain([
     testCase(MissingDocsRuleConfigurationTests.allTests),
     testCase(MissingDocsRuleTests.allTests),
     testCase(ModifierOrderTests.allTests),
+    testCase(MultilineArgumentsBracketsRuleTests.allTests),
     testCase(MultilineArgumentsRuleTests.allTests),
     testCase(MultilineFunctionChainsRuleTests.allTests),
+    testCase(MultilineLiteralBracketsRuleTests.allTests),
+    testCase(MultilineParametersBracketsRuleTests.allTests),
     testCase(MultilineParametersRuleTests.allTests),
     testCase(MultipleClosuresWithTrailingClosureRuleTests.allTests),
     testCase(NestingRuleTests.allTests),